frontend: Autofocuses on newly added tag input
